backend/requirements.txt
# ── Reddit API client ──────────────────────────────────────────────────────────
praw>=7.7.1               # Python Reddit API Wrapper — primary data-fetching library
prawcore>=2.4.0           # Low-level HTTP client used internally by PRAW; pinned for
                          # stable rate-limit exception handling

# ── LLM / AI orchestration ────────────────────────────────────────────────────
openai>=1.30.0            # OpenAI Python SDK (v1+ async-compatible)
langchain>=0.2.0          # Core LangChain abstractions (chains, prompts, callbacks)
langchain-openai>=0.1.8   # LangChain ↔ OpenAI integration (ChatOpenAI, embeddings)
langgraph>=0.1.0          # Graph-based agent orchestration built on LangChain

# ── Data validation ───────────────────────────────────────────────────────────
pydantic>=2.7.0           # Runtime type validation for Thread / Comment models
pydantic-settings>=2.2.0  # Settings management on top of Pydantic v2

# ── Configuration & environment ───────────────────────────────────────────────
python-dotenv>=1.0.1      # Loads .env files into os.environ

# ── Web API layer ─────────────────────────────────────────────────────────────
fastapi>=0.111.0          # Modern async web framework for the REST API
uvicorn[standard]>=0.29.0 # ASGI server with uvloop + httptools for performance
httpx>=0.27.0             # Async HTTP client (used by FastAPI test client & agents)

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.2.0             # Test runner
pytest-asyncio>=0.23.6    # Async test support (needed for FastAPI + HTTPX tests)
